mailer-service/pyproject.toml
[tool.poetry]
name = "mailer"
version = "0.1.0"
description = "Mailer service for sending emails via RabbitMQ"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aio-pika = "^8.0.0"
aiosmtplib = "^1.1.4"
python-dotenv = "^1.0.0"
asyncio = "^3.4.3"
pika = "^1.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mailer = "mailer.mailer:main"